--- src/components/Button/addbutton.js
@@ -1,5 +1,0 @@
-import StyledButton from './styledbutton.js';
-
-export default function AddButton() {
-	return <StyledButton>Add</StyledButton>;
-}
