--- src/components/Button/backbutton.js
@@ -1,9 +1,0 @@
-import {useNavigate} from 'react-router-dom';
-
-import StyledButton from './styledbutton.js';
-
-export default function BackButton() {
-	const navigate = useNavigate();
-
-	return <StyledButton onClick={() => navigate('/')}>Go Back</StyledButton>;
-}
