--- src/components/Button/deletebutton.js
@@ -1,9 +1,0 @@
-import StyledIcon from './styledicon';
-
-export default function DeleteButton({onClick}) {
-	return (
-		<StyledIcon onClick={onClick}>
-			<i className="fa-solid fa-trash"></i>
-		</StyledIcon>
-	);
-}
