--- src/components/Button/savebutton.js
@@ -1,5 +1,0 @@
-import StyledButton from './styledbutton.js';
-
-export default function SaveButton() {
-	return <StyledButton>Save</StyledButton>;
-}
