+++ src/components/CreateForm/index.js
@@ -1,7 +1,7 @@
 import {useState} from 'react';
 
 import useStore from '../../hooks/useStore.js';
-import AddButton from '../Button/addbutton.js';
+import Button from '../Button/Button.js';
 
 import Radio from './radio.js';
 import StyledForm from './styledform.js';
@@ -29,7 +29,7 @@ export default function CreateForm() {
 				<h2>Plan your next meal!</h2>
 				<TextInput newMeal={newMeal} setNewMeal={setNewMeal} />
 				<Radio newMeal={newMeal} setNewMeal={setNewMeal} />
-				<AddButton type="submit" />
+				<Button buttonMode={'submit'}>Add</Button>
 			</StyledForm>
 		</>
 	);
