+++ src/components/MealCard/index.js
@@ -1,5 +1,7 @@
+import {useNavigate} from 'react-router-dom';
+
 import useStore from '../../hooks/useStore.js';
-import CardButton from '../Button/cardbutton.js';
+import Button from '../Button/Button.js';
 
 import MealCardContainer from './styled.js';
 
@@ -7,11 +9,15 @@ export default function MealCard({id}) {
 	const allMeals = useStore(state => state.allMeals);
 	const meal = allMeals.find(meal => meal.id === id);
 
+	const navigate = useNavigate();
+
 	return (
 		<MealCardContainer>
 			<h2>{meal.title}</h2>
 			<p>{meal.category}</p>
-			<CardButton id={meal.id} />
+			<Button functionToClick={navigate} parameterToClick={'/' + id}>
+				Details
+			</Button>
 		</MealCardContainer>
 	);
 }
